Re-write commit history mes.Rd
There was some fear that my function (now called solar_daily) would confuse folks looking for the original research paper and code. As such, I've purged references to the old function name to minimize the chances of a search engine leading folks here instead of to the original author's work.